Move WorkflowNode to Node
