+++ src/Activity/Activity.php
@@ -2,7 +2,7 @@
 
 namespace Phlow\Activity;
 
-use Phlow\Model\WorkflowNode;
+use Phlow\Node\WorkflowNode;
 
 /**
  * Interface Activity
