+++ src/Event/EndEvent.php
@@ -3,7 +3,7 @@
 namespace Phlow\Event;
 
 use Phlow\Model\RenderableObject;
-use Phlow\Model\WorkflowNodeTrait;
+use Phlow\Node\WorkflowNodeTrait;
 
 /**
  * Class EndEvent
