+++ src/Event/ErrorEvent.php
@@ -5,7 +5,7 @@ namespace Phlow\Event;
 use Phlow\Model\ExceptionHandlerNode;
 use Phlow\Model\ExceptionHandlerTrait;
 use Phlow\Model\RenderableObject;
-use Phlow\Model\WorkflowNodeTrait;
+use Phlow\Node\WorkflowNodeTrait;
 
 /**
  * Class ErrorEvent
